entity created and fixes
Added birth date, gender, and residence fields.

Renamed the enum Active to State.

In TicketMachine, added a "type" field and set the identificationCode to be unique.

Added surname to the Seller entity.

Created DAO classes for TicketSeller and TicketMachine.

Renamed the save and findById methods in VehiclesDAO and VehiclesMaintenanceDAO to keep naming consistent across the project.